requirements-prod.txt
# Core dependencies
langgraph>=0.2.0
langchain-google-genai>=0.1.0
pypdf>=4.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# FastAPI and web server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Production optimizations
gunicorn>=21.0.0
